chore: Clean up types
