+++ src/Icon/Icon.tsx
@@ -1,11 +1,10 @@
-import PropTypes from 'prop-types';
 import { css } from 'styled-components';
 import { createComponent } from '../utils';
 
 interface IconProps {
   name: string;
   size?: number;
-  color?: string;
+  color: string;
   onClick?: () => void;
   disabled?: boolean;
 }
@@ -16,8 +15,8 @@ const Icon = createComponent<IconProps, 'i'>({
   props: ({ name }) => ({
     className: `mdi mdi-${name}`,
   }),
-  style: ({ theme, size, color, disabled, onClick }: any) => {
-    const colorFromTheme = theme.colors[color];
+  style: ({ theme, size, color, disabled, onClick }) => {
+    const colorFromTheme: string = theme.colors[color];
     const resolvedColor = colorFromTheme || color;
 
     return css`
@@ -37,12 +36,5 @@ const Icon = createComponent<IconProps, 'i'>({
     `;
   },
 });
-
-Icon.propTypes = {
-  name: PropTypes.string.isRequired,
-  size: PropTypes.number,
-  color: PropTypes.string,
-  onClick: PropTypes.func,
-};
 
 export default Icon;
