+++ src/defaultBadgeVariants.ts
@@ -1,29 +1,32 @@
-import { defaultThemeColors } from './defaultThemeColors';
+import { VariantObject } from './types';
 
-// Note - Typing is generated by this constant
-export const defaultBadgeVariants = {
-  primary: {
-    backgroundColor: defaultThemeColors.primaryLightest,
-    color: defaultThemeColors.primaryDark,
-  },
-  success: {
-    backgroundColor: defaultThemeColors.greenLightest,
-    color: defaultThemeColors.greenDark,
-  },
-  danger: {
-    backgroundColor: defaultThemeColors.redLightest,
-    color: defaultThemeColors.redDark,
-  },
-  warning: {
-    backgroundColor: defaultThemeColors.orangeLightest,
-    color: defaultThemeColors.orangeDark,
-  },
-  info: {
-    backgroundColor: defaultThemeColors.blueLightest,
-    color: defaultThemeColors.blueDark,
-  },
-  grey: {
-    backgroundColor: defaultThemeColors.greyLight,
-    color: defaultThemeColors.greyDarker,
-  },
-} as const;
+export const getDefaultBadgeVariants = (colors: Record<string, string>): VariantObject =>
+  ({
+    primary: {
+      backgroundColor: colors.primaryLightest,
+      color: colors.primaryDark,
+    },
+    success: {
+      backgroundColor: colors.greenLightest,
+      color: colors.greenDark,
+      '&:before': {
+        color: colors.green,
+      },
+    },
+    danger: {
+      backgroundColor: colors.redLightest,
+      color: colors.redDark,
+    },
+    warning: {
+      backgroundColor: colors.orangeLightest,
+      color: colors.orangeDark,
+    },
+    info: {
+      backgroundColor: colors.blueLightest,
+      color: colors.blueDark,
+    },
+    grey: {
+      backgroundColor: colors.greyLight,
+      color: colors.greyDarker,
+    },
+  } as const);
