+++ src/defaultBreakpoints.ts
@@ -1,5 +1,3 @@
-// Note - Typing is generated by this constant
-
 export const defaultBreakpoints = {
   xs: '400px',
   sm: '600px',
