+++ src/defaultButtonVariants.ts
@@ -1,109 +1,109 @@
-import { defaultThemeColors } from './defaultThemeColors';
+import { VariantObject } from './types';
 
-// Note - Typing is generated by this constant
-export const defaultButtonVariants = {
-  primary: {
-    backgroundColor: defaultThemeColors.primary,
-    color: defaultThemeColors.white,
-    '&:hover': {
-      backgroundColor: defaultThemeColors.primaryLight,
+export const getDefaultButtonVariants = (colors: Record<string, string>): VariantObject =>
+  ({
+    primary: {
+      backgroundColor: colors.primary,
+      color: colors.white,
+      '&:hover': {
+        backgroundColor: colors.primaryLight,
+      },
+      '&:active': {
+        backgroundColor: colors.primaryDark,
+      },
+      '&:disabled': {
+        backgroundColor: colors.primaryLightest,
+      },
     },
-    '&:active': {
-      backgroundColor: defaultThemeColors.primaryDark,
+    primaryText: {
+      backgroundColor: 'transparent',
+      borderColor: 'transparent',
+      color: colors.primary,
+      '&:hover': {
+        color: colors.primaryDark,
+      },
     },
-    '&:disabled': {
-      backgroundColor: defaultThemeColors.primaryLightest,
+    secondary: {
+      backgroundColor: 'transparent',
+      borderColor: colors.primary,
+      color: colors.primary,
+      '&:hover': {
+        backgroundColor: colors.primary,
+        color: colors.white,
+      },
+      '&:active': {
+        backgroundColor: colors.primaryDark,
+        color: colors.white,
+      },
+      '&:disabled': {
+        borderColor: colors.primaryLightest,
+        color: colors.primaryLightest,
+      },
     },
-  },
-  primaryText: {
-    backgroundColor: 'transparent',
-    borderColor: 'transparent',
-    color: defaultThemeColors.primary,
-    '&:hover': {
-      color: defaultThemeColors.primaryDark,
+    grey: {
+      backgroundColor: colors.white,
+      borderColor: colors.grey,
+      color: colors.greyDarkest,
+      '&:hover': {
+        borderColor: colors.greyDark,
+      },
+      '&:active': {
+        backgroundColor: colors.greyLight,
+      },
+      '&:disabled': {
+        borderColor: colors.grey,
+        color: colors.grey,
+      },
     },
-  },
-  secondary: {
-    backgroundColor: 'transparent',
-    borderColor: defaultThemeColors.primary,
-    color: defaultThemeColors.primary,
-    '&:hover': {
-      backgroundColor: defaultThemeColors.primary,
-      color: defaultThemeColors.white,
+    greyText: {
+      backgroundColor: 'transparent',
+      borderColor: 'transparent',
+      color: colors.grey,
+      '&:hover': {
+        color: colors.greyDark,
+      },
     },
-    '&:active': {
-      backgroundColor: defaultThemeColors.primaryDark,
-      color: defaultThemeColors.white,
+    success: {
+      color: colors.white,
+      backgroundColor: colors.secondary,
+      '&:hover': {
+        backgroundColor: colors.secondaryLight,
+      },
+      '&:active': {
+        backgroundColor: colors.secondaryDark,
+      },
+      '&:disabled': {
+        backgroundColor: colors.secondaryLightest,
+      },
     },
-    '&:disabled': {
-      borderColor: defaultThemeColors.primaryLightest,
-      color: defaultThemeColors.primaryLightest,
+    warning: {
+      backgroundColor: colors.orange,
+      color: colors.white,
+      '&:hover': {
+        backgroundColor: colors.orangeLight,
+      },
+      '&:active': {
+        backgroundColor: colors.orangeDark,
+      },
+      '&:disabled': {
+        backgroundColor: colors.orangeLightest,
+      },
     },
-  },
-  grey: {
-    backgroundColor: defaultThemeColors.white,
-    borderColor: defaultThemeColors.grey,
-    color: defaultThemeColors.greyDarkest,
-    '&:hover': {
-      borderColor: defaultThemeColors.greyDark,
+    danger: {
+      backgroundColor: colors.red,
+      color: colors.white,
+      '&:hover': {
+        backgroundColor: colors.redLight,
+      },
+      '&:active': {
+        backgroundColor: colors.redDark,
+      },
+      '&:disabled': {
+        backgroundColor: colors.redLightest,
+      },
     },
-    '&:active': {
-      backgroundColor: defaultThemeColors.greyLight,
+    info: {
+      backgroundColor: colors.blue,
+      color: colors.white,
     },
-    '&:disabled': {
-      borderColor: defaultThemeColors.grey,
-      color: defaultThemeColors.grey,
-    },
-  },
-  greyText: {
-    backgroundColor: 'transparent',
-    borderColor: 'transparent',
-    color: defaultThemeColors.grey,
-    '&:hover': {
-      color: defaultThemeColors.greyDark,
-    },
-  },
-  success: {
-    color: defaultThemeColors.white,
-    backgroundColor: defaultThemeColors.secondary,
-    '&:hover': {
-      backgroundColor: defaultThemeColors.secondaryLight,
-    },
-    '&:active': {
-      backgroundColor: defaultThemeColors.secondaryDark,
-    },
-    '&:disabled': {
-      backgroundColor: defaultThemeColors.secondaryLightest,
-    },
-  },
-  warning: {
-    backgroundColor: defaultThemeColors.orange,
-    color: defaultThemeColors.white,
-    '&:hover': {
-      backgroundColor: defaultThemeColors.orangeLight,
-    },
-    '&:active': {
-      backgroundColor: defaultThemeColors.orangeDark,
-    },
-    '&:disabled': {
-      backgroundColor: defaultThemeColors.orangeLightest,
-    },
-  },
-  danger: {
-    backgroundColor: defaultThemeColors.red,
-    color: defaultThemeColors.white,
-    '&:hover': {
-      backgroundColor: defaultThemeColors.redLight,
-    },
-    '&:active': {
-      backgroundColor: defaultThemeColors.redDark,
-    },
-    '&:disabled': {
-      backgroundColor: defaultThemeColors.redLightest,
-    },
-  },
-  info: {
-    backgroundColor: defaultThemeColors.blue,
-    color: defaultThemeColors.white,
-  },
-} as const;
+  } as const);
