+++ src/defaultThemeColors.ts
@@ -1,4 +1,3 @@
-// Note - Typing is generated by this constant
 export const defaultThemeColors = {
   default: '#494D55',
   black: '#1B202B',
