+++ src/defaultThemeSizes.ts
@@ -1,4 +1,3 @@
-// Note - Typing is generated by this constant
 export const defaultThemeSizes = {
   Button: {
     sm: {
