--- src/types/BadgeVariants.ts
@@ -1,8 +1,0 @@
-import { defaultBadgeVariants } from '../defaultBadgeVariants';
-import { ExtendedCSSProperties } from './ExtendedCSSProperties';
-
-type Transform<T> = {
-  [P in keyof T]: ExtendedCSSProperties;
-};
-
-export interface BadgeVariants extends Transform<typeof defaultBadgeVariants> {}
