--- src/types/ButtonVariants.ts
@@ -1,8 +1,0 @@
-import { defaultButtonVariants } from '../defaultButtonVariants';
-import { ExtendedCSSProperties } from './ExtendedCSSProperties';
-
-type Transform<T> = {
-  [P in keyof T]: ExtendedCSSProperties;
-};
-
-export interface ButtonVariants extends Transform<typeof defaultButtonVariants> {}
