--- src/types/ThemeBreakpoints.ts
@@ -1,7 +1,0 @@
-import { defaultBreakpoints } from '../defaultBreakpoints';
-
-type Transform<T> = {
-  [P in keyof T]: string;
-};
-
-export interface ThemeBreakpoints extends Transform<typeof defaultBreakpoints> {}
