--- src/types/ThemeColors.ts
@@ -1,7 +1,0 @@
-import { defaultThemeColors } from 'src/defaultThemeColors';
-
-type Transform<T> = {
-  [P in keyof T]: string;
-};
-
-export interface ThemeColors extends Transform<typeof defaultThemeColors> {}
