--- src/types/ThemeSizes.ts
@@ -1,8 +1,0 @@
-import { defaultThemeSizes } from '../defaultThemeSizes';
-import { ThemeSize } from './ThemeSize';
-
-type Transform<T> = {
-  [P in keyof T]: ThemeSize;
-};
-
-export interface ThemeSizes extends Transform<typeof defaultThemeSizes> {}
