--- src/types/ThemeTypography.ts
@@ -1,6 +1,0 @@
-export interface ThemeTypography {
-  fontSize: number;
-  color: string;
-  bodyFontFamily: string;
-  headerFontFamily: string;
-}
