--- src/types/ThemeVariants.ts
@@ -1,8 +1,0 @@
-import { ButtonVariants } from './ButtonVariants';
-import { BadgeVariants } from './BadgeVariants';
-
-export interface ThemeVariants {
-  Alert: BadgeVariants;
-  Badge: BadgeVariants;
-  Button: ButtonVariants;
-}
