+++ src/utils.ts
@@ -1,6 +1,6 @@
 import kebabCase from 'lodash/kebabCase';
 import get from 'lodash/get';
-import styled from 'styled-components';
+import styled, { DefaultTheme, SimpleInterpolation } from 'styled-components';
 
 export const themeGet = (lookup: any, fallback?: any) => ({ theme }: any = {}) => get(theme, lookup, fallback);
 
@@ -34,12 +34,12 @@ const getComponentClassName = (
   name: string
 ) => `${className || ''} ${classPrefix}-${name} ${variant ? `${classPrefix}-${name}-${variant}` : ''}`.trim();
 
-interface CreateComponentProps {
+interface CreateComponentProps<Props> {
   name: string;
   tag?: keyof JSX.IntrinsicElements;
   as?: React.ComponentType<any>;
-  style?: any;
-  props?: (props: any) => any;
+  style?: readonly SimpleInterpolation[] | ((props: Props & { theme: DefaultTheme }) => readonly SimpleInterpolation[]);
+  props?: (props: Props) => Record<string, string | number | boolean>;
 }
 
 export const createComponent = <T extends object, O extends keyof JSX.IntrinsicElements | React.ComponentType<any>>({
@@ -48,10 +48,10 @@ export const createComponent = <T extends object, O extends keyof JSX.IntrinsicE
   as,
   style,
   props: getBaseProps = () => ({}),
-}: CreateComponentProps) => {
+}: CreateComponentProps<T>) => {
   const component = styled<O>((as || tag) as any);
 
-  return component.attrs<T>((props: any) => {
+  return component.attrs<T>(props => {
     const baseProps = getBaseProps(props);
     const finalProps = {
       ...baseProps,
@@ -66,7 +66,7 @@ export const createComponent = <T extends object, O extends keyof JSX.IntrinsicE
   })<T>`
   ${style}
   ${getComponentStyle(name)}
-  ${(p: any) => getVariantStyles(name, p.variant)}
+  ${p => getVariantStyles(name, p.variant)}
   ${({ styles = {} }) => styles[name] || {}}
 `;
 };
